+++ src/auth/dto/create.buyer.dto.ts
@@ -1,8 +1,7 @@
 import { Transform } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmptyNumber } from 'src/decorators/is-not-empty-number.decorator';
-import { IsNotEmptyString } from 'src/decorators/is-not-empty-string.decorator';
 import { BuyerEntity } from 'src/entities/buyer.entity';
+import { IsNotEmptyString, IsNotEmptyNumber } from 'src/util/decorator/validate.decorater';
 import { AuthCommonDto } from './auth.common.dto';
 
 export class CreateBuyerDto extends AuthCommonDto implements Pick<BuyerEntity, 'name' | 'gender' | 'age' | 'phone'> {
