--- src/decorators/is-optional-boolean.decorator.ts
@@ -1,18 +1,0 @@
-import { Transform } from 'class-transformer';
-import { IsBoolean, IsOptional } from 'class-validator';
-import { applyDecorators } from '@nestjs/common';
-
-const optionalBooleanMapper = new Map([
-  ['undefined', undefined],
-  ['true', true],
-  ['false', false],
-]);
-
-export const ParseOptionalBoolean = () =>
-  Transform((params) => {
-    return optionalBooleanMapper.get(String(params.value));
-  });
-
-export function IsOptionalBoolean() {
-  return applyDecorators(IsOptional(), ParseOptionalBoolean(), IsBoolean());
-}
