+++ src/entities/dtos/get-product-list-pagination.dto.ts
@@ -1,6 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsOptionalNumber } from 'src/decorators/is-optional-number.decorator';
-import { IsOptionalString } from 'src/decorators/is-optional-string.decorator';
+import { IsOptionalString, IsOptionalNumber } from 'src/util/decorator/validate.decorater';
 import { PaginationDto } from './pagination.dto';
 
 export class GetProductListPaginationDto extends PaginationDto {
