+++ src/seller/dto/create.product.options.dto.ts
@@ -1,8 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmptyBoolean } from 'src/decorators/is-not-empty-boolean.decorator';
-import { IsNotEmptyNumber } from 'src/decorators/is-not-empty-number.decorator';
-import { IsNotEmptyString } from 'src/decorators/is-not-empty-string.decorator';
 import { ProductRequiredOptionEntity } from 'src/entities/product-required-option.entity';
+import { IsNotEmptyString, IsNotEmptyNumber, IsNotEmptyBoolean } from 'src/util/decorator/validate.decorater';
 
 export class CreateProductOptionsDto implements Pick<ProductRequiredOptionEntity, 'name' | 'price' | 'isSale'> {
   @ApiProperty({ type: String, description: '옵션 이름', required: true })
