+++ src/seller/dto/is.require.option.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsNotEmptyBoolean } from 'src/decorators/is-not-empty-boolean.decorator';
+import { IsNotEmptyBoolean } from 'src/util/decorator/validate.decorater';
 
 export class IsRequireOptionDto {
   @ApiProperty({ type: Boolean, description: '필수옵션 여부', required: true })
